requirements.txt
pillow
numpy
jinja2
pydantic
opencv-python
